"Documentation" release

+++ application/Locodio/Application/Command/Model/DeleteDomainModel/DeleteDomainModelHandler.php
@@ -2,6 +2,7 @@
 
 namespace App\Locodio\Application\Command\Model\DeleteDomainModel;
 
+use App\Locodio\Application\Command\Model\ModelFinalChecker;
 use App\Locodio\Domain\Model\Model\DomainModelRepository;
 use App\Locodio\Domain\Model\Model\AttributeRepository;
 use App\Locodio\Domain\Model\Model\AssociationRepository;
@@ -26,6 +27,10 @@ class DeleteDomainModelHandler
     public function go(DeleteDomainModel $command): bool
     {
         $domainModel = $this->domainModelRepo->getById($command->getId());
+        if (ModelFinalChecker::isFinalState($domainModel->getDocumentor())) {
+            return false;
+        }
+
         foreach ($domainModel->getAttributes() as $attribute) {
             $this->attributeRepo->delete($attribute);
         }
@@ -33,6 +38,7 @@ class DeleteDomainModelHandler
             $this->associationRepo->delete($association);
         }
         $this->domainModelRepo->delete($domainModel);
+
         return true;
     }
 }
