+++ application/Locodio/Application/Command/Model/DeleteEnumOption/DeleteEnumOptionHandler.php
@@ -2,6 +2,7 @@
 
 namespace App\Locodio\Application\Command\Model\DeleteEnumOption;
 
+use App\Locodio\Application\Command\Model\ModelFinalChecker;
 use App\Locodio\Domain\Model\Model\EnumOptionRepository;
 
 class DeleteEnumOptionHandler
@@ -21,7 +22,12 @@ class DeleteEnumOptionHandler
     public function go(DeleteEnumOption$command): bool
     {
         $enumOption =  $this->enumOptionRepo->getById($command->getId());
+        if (ModelFinalChecker::isFinalState($enumOption->getEnum()->getDocumentor())) {
+            return false;
+        }
+
         $this->enumOptionRepo->delete($enumOption);
+
         return true;
     }
 }
